+++ nameable.rb
@@ -1,0 +1,6 @@
+# Interface
+class Nameable
+  def correct_name
+    raise NotImplementedError, 'This method must be implemented in the child class.'
+  end
+end
